Refactor JSON parsing in translation process
The custom JSON extraction and validation method has been replaced with the standard 'JSON.parse()' for simpler and more direct parsing. The returned errors and some console logs were also removed to clean up the code. The ts-ignore directive was added to suppress TypeScript warnings for a certain line of code.